backend/requirements.txt
pyjwt==2.8.0
django==5.0.2
pydantic==2.6.1
psycopg==3.1.18
django-ninja==1.1.0
cryptography==42.0.2
django-ninja-jwt==5.3.0
django-model-utils==4.4.0
django-ninja-extra==0.20.4
django-admin-tailwind==0.0.31

# For testing only
pylint==3.0.4
coverage==7.4.3
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0